Use oplog hack only for forward cursor
1. It's not obvious if backward cursor benefits from this hack.
2. Backward cursor hack breaks the rollback logic. Mongo uses backward
cursor to do capped collection rollback and truncation by going backwards
and observing the first entry earlier than the given one.
So it's expecting the cursor to show ALL documents, not only committed
ones.
If some documents are hidden, the truncation may not happen as it's
considered a situation when there's nothing to rollback and truncate.
Backward cursors' behavior differs from what's expected of forward cursors.